Let daemon write to file, if it has ownership, add --file parameter
Implement open/close messages.
