+++ rust/src/types.rs
@@ -1,7 +1,6 @@
 #![allow(dead_code)] // TODO: consider the dead code.
 use automerge::PatchAction;
 use operational_transform::{Operation as OTOperation, OperationSeq};
-use serde_json::Value as JSONValue;
 
 #[derive(Debug, Clone, PartialEq, Default)]
 pub struct TextDelta(pub Vec<TextOp>);
@@ -49,8 +48,8 @@ pub struct EditorTextOp {
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct Range {
-    anchor: Position,
-    head: Position,
+    pub anchor: Position,
+    pub head: Position,
 }
 
 impl Range {
@@ -273,122 +272,6 @@ impl From<TextDelta> for EditorTextDelta {
     }
 }
 
-// TODO: Write this in a nicer way...
-impl TryFrom<JSONValue> for RevisionedEditorTextDelta {
-    type Error = anyhow::Error;
-    fn try_from(json: JSONValue) -> Result<Self, Self::Error> {
-        match json {
-            serde_json::Value::Object(map) => {
-                if let Some(serde_json::Value::String(method)) = map.get("method") {
-                    match method.as_str() {
-                        "insert" => {
-                            if let Some(serde_json::Value::Array(array)) = map.get("params") {
-                                if let Some(serde_json::Value::Number(revision)) = array.get(1) {
-                                    if let Some(serde_json::Value::Number(position)) = array.get(2)
-                                    {
-                                        if let Some(serde_json::Value::String(text)) = array.get(3)
-                                        {
-                                            let revision = revision
-                                                .as_u64()
-                                                .expect("Failed to parse revision");
-                                            let position = position
-                                                .as_u64()
-                                                .expect("Failed to parse position");
-                                            let text = text.as_str().to_string();
-                                            let op = EditorTextOp {
-                                                range: Range {
-                                                    anchor: position as usize,
-                                                    head: position as usize,
-                                                },
-                                                replacement: text,
-                                            };
-                                            let delta = EditorTextDelta(vec![op]);
-                                            Ok(RevisionedEditorTextDelta {
-                                                revision: revision as usize,
-                                                delta,
-                                            })
-                                        } else {
-                                            Err(anyhow::anyhow!(
-                                                "Could not find text param in position #3"
-                                            ))
-                                        }
-                                    } else {
-                                        Err(anyhow::anyhow!(
-                                            "Could not find position param in position #2"
-                                        ))
-                                    }
-                                } else {
-                                    Err(anyhow::anyhow!(
-                                        "Could not find revision param in position #1"
-                                    ))
-                                }
-                            } else {
-                                Err(anyhow::anyhow!("Could not find params for insert method"))
-                            }
-                        }
-                        "delete" => {
-                            if let Some(serde_json::Value::Array(array)) = map.get("params") {
-                                if let Some(serde_json::Value::Number(revision)) = array.get(1) {
-                                    if let Some(serde_json::Value::Number(position)) = array.get(2)
-                                    {
-                                        if let Some(serde_json::Value::Number(length)) =
-                                            array.get(3)
-                                        {
-                                            let revision = revision
-                                                .as_u64()
-                                                .expect("Failed to parse revision");
-                                            let position = position
-                                                .as_u64()
-                                                .expect("Failed to parse position");
-
-                                            let length =
-                                                length.as_u64().expect("Failed to parse length");
-
-                                            let op = EditorTextOp {
-                                                range: Range {
-                                                    anchor: position as usize,
-                                                    head: position as usize + length as usize,
-                                                },
-                                                replacement: "".to_string(),
-                                            };
-                                            let delta = EditorTextDelta(vec![op]);
-                                            Ok(RevisionedEditorTextDelta {
-                                                revision: revision as usize,
-                                                delta,
-                                            })
-                                        } else {
-                                            Err(anyhow::anyhow!(
-                                                "Could not find length param in position #3"
-                                            ))
-                                        }
-                                    } else {
-                                        Err(anyhow::anyhow!(
-                                            "Could not find position param in position #2"
-                                        ))
-                                    }
-                                } else {
-                                    Err(anyhow::anyhow!("Could not find params for delete method"))
-                                }
-                            } else {
-                                Err(anyhow::anyhow!(
-                                    "Could not find revision param in position #1"
-                                ))
-                            }
-                        }
-                        _ => Err(anyhow::anyhow!("Unknown JSON method: {}", method)),
-                    }
-                } else {
-                    Err(anyhow::anyhow!("Could not find method in JSON message"))
-                }
-            }
-            _ => Err(anyhow::anyhow!(
-                "JSON message is not an object: {:#?}",
-                json
-            )),
-        }
-    }
-}
-
 #[cfg(test)]
 pub mod factories {
     use super::*;
@@ -462,38 +345,5 @@ mod tests {
         expected.insert("short");
         expected.insert("ong");
         assert_eq!(delta, expected);
-    }
-
-    #[test]
-    fn json_to_revisioned_editor_text_delta() {
-        let json = serde_json::json!({
-            "method": "insert",
-            "params": ["", 0, 1, "a"]
-        });
-
-        let delta = RevisionedEditorTextDelta::try_from(json).unwrap();
-        assert_eq!(delta.revision, 0);
-        assert_eq!(
-            delta.delta,
-            EditorTextDelta(vec![EditorTextOp {
-                range: Range { anchor: 1, head: 1 },
-                replacement: "a".to_string(),
-            }])
-        );
-
-        let json = serde_json::json!({
-            "method": "delete",
-            "params": ["", 2, 1, 3]
-        });
-
-        let delta = RevisionedEditorTextDelta::try_from(json).unwrap();
-        assert_eq!(delta.revision, 2);
-        assert_eq!(
-            delta.delta,
-            EditorTextDelta(vec![EditorTextOp {
-                range: Range { anchor: 1, head: 4 },
-                replacement: "".to_string(),
-            }])
-        );
     }
 }
